requirements.txt
# checker
asyncio
beautifulsoup4
lxml
requests
unidecode

# bot
python-telegram-bot
unidecode
